Enforcing boolean operator precedence
The operator precedence is now:

1. NOT/!
2. AND/&&
3. OR/||

Where 1 has the tightest binding, and 3 has the lowest. Terms separated
by an implicit operator `a b` will be treated similarly as an `OR` with
a *slightly* lower precedence.

The AST tests needed tweaked a bit to shift order around. The AST itself
can be cleaned up a fair bit from here, but this was a big enough
change.